requirements.txt
flask
pymongo
python-gnupg